remove helpers for exporting variables globally
